feat: add hbox container and demo
