+++ app/index.html
@@ -11,6 +11,7 @@
     <!-- Elemental UI components -->
     <script type="module" src="/components/basic/button.js"></script>
     <script type="module" src="/components/container/vbox.js"></script>
+    <script type="module" src="/components/container/hbox.js"></script>
     
     <script type="module" src="/src/main.jsx"></script>
   </body>
