+++ components/index.js
@@ -1,2 +1,3 @@
 export * from './basic/button.js';
 export * from './container/vbox.js';
+export * from './container/hbox.js';
